# Conflicts:
#	app/src/main/java/com/multimoney/multimoney/presentation/navigation/Screen.kt

+++ domain/src/main/java/com/multimoney/domain/interaction/accountsmart/QueryStepByStepUseCase.kt
@@ -8,6 +8,6 @@ interface QueryStepByStepUseCase {
     suspend operator fun invoke(
         user: String,
         idBrand: Int,
-        idRequest: Int,
+        idRequest: Long
     ): Flow<MultimoneyResult<StepByStep?>>
 }
